refactor: add agent directory manager
